v 0.4

+++ js/main.js
@@ -1,15 +1,18 @@
-//تحميل البيانات
+// load data
 let all_data = {};
 
-fetch("../data/data.json")
+const existingData = localStorage.getItem("all_data");
+
+if (existingData) {
+    all_data = JSON.parse(existingData);
+} else {
+    fetch("../data/data.json")
     .then(res => res.json())
     .then(data => {
         localStorage.setItem("all_data", JSON.stringify(data));
-        all_data = data;
-
-        all_data = JSON.parse(localStorage.getItem("all_data")) || [];
+        
     })
     .catch(err => console.error("error in loading data", err));
     
-
-    //-------------------------------------------------------------------
+}
+//-------------------------------------------------------------------
